Tighten TLS and save ALL pem files
This tightens up the TLS config generated a little bit. It should now
get a near perfect qualys score.

Before we only copied the account key, the encryption key and the
full chain file. This was causeing problems because when acme-client
went to validate the cert, it would try to validate cert.pem which we
were not saving/restoring. This patch adds cert.pem and chain.pem to
the saved/restored files.