Resolve issue with blue/gree deployment
I seperated out pulling certs from consul from the part that runs with a
lock. The reason for this is lock ends up preventing pulling the cert from
consul for half a day while it's waiting to aquire the lock (possibly
longer, if there are more then just two nginx servers). This should allow
nginx to start serving encrypted traffic even with another nginx server
holding the lock.

I also added the -e flag to acme-client so it can reissue the cert of the
SAN (Subject Alternate Name) changes.