Restart nginx if fullchain.pem changes
I added the writing of /etc/nginx/ssl/configured (which triggers nginx to
reload) to after (and conditional on) pulling nginx/acme/fullchain.pem
because otherwise it waits for the lock release to reload the nginx
config. Again, this was causing a half (or more) day delay on servers
past the first serving encrypted traffic.

Because of this change, I made the restart on update by acme-client
dependent on acme-client actually making a change. This should reduce
unnessicary restarts. (Always doing it was just silliness on my part,
but it was only twice-a-day silliness)