Fix signal handling in acme-manage
Signal handling in shell scripts is ugly, so bear with me.

In order for acme-manage to properly handle the signal from update-domains,
it needs to background consul lock, save it's PID, and wait for it to
finish. We also exec the sleep in do_client, so when consul lock passes a
signal to it's child, it's really passing it to sleep which has taken over
the script's PID. This means the instance of the script running do_client
is already gone once the sleep happens, the sleep terminates because it
get's the signal consul lock intended for do_client, consul lock gets a
signal and properly shuts down, and the acme-mangage client can remove the
PID file and exit normally.