Alter acme to use a shared volume
A shared volume makes grabbing the certs MUCH easier. We don't have to
worry about locking semantics of the storage, because every time we
write one of the files here, we have a consul lock.

The dhparam.pem is also now stored here, and the functions which were
handled by dhparam-init and update-domains are now handled by acme-manage.