Set /etc/ssl/acme/configured to 0 by default
This file changes every time we update the ssl config. When it changes,
the nginx.conf is regenerated and nginx is restarted. For the creation of
the nginx.conf, it has to have a value in it, and if it's NOT configured
that value has to be an integer which evaluates to false (which is to say
0). So if the file doesn't exist the master sets it to 0 so nginx can load
in non-ssl mode (this will change as soon as acme-client gets the keys).

Also fixed a bug that would cause dhparams to recreate once for every
host.